FLTS-4452: Add TAB-completion support to frija
TAB-support added to support completion of sub-commands and their
options. This works by letting the frija command (actually Bash
function) source the intended frija sub-command script. This sets
global variables with options that can be completed as well
as (re-)defines global Bash-functions used in the completion process.

The common Frija-completion machine then operates a state machine
based on these values.

Due to that the sub-command scripts are sourced, detection for this is
added to these scripts so the source command return before the
complete script is sourced. Otherwise the complete sub-command would
be executed which is not the intent of the completion operation.