FLTS-4452: Add completion support
The bootstrap function has been reworked due to other changes made to
how the frija commands work. Now it is a separate script placed in the
root of the metadata_tools_01 repo simply called "bootstrap" that in
turn reuses code from the frija-init sub-command.

Some of the changes introduced are to make it possible for the
bootstrap script to source the frija-init sub-command to make
the initialize() function it defines available to it.

Lots of minor fixes and cleanup.

* Generate Directory.Build.props file for MSBuild-based repos

* Create installation archive via two stages (prepare and archive) for
  MSBuild

* When fetching repos add option to also update all local branches at
  the same time

* Tagging of repos with where the tag is a combination of a version
  and short SHA for the commit. Also ensure tag is attached to a
  commit matching the short SHA in the tag name.