FLTS-4452: Make clone command switch to branch
Concept of "latest" is replaced with "floating" as this is more clear
what it is about; repo is not "locked" to a specific version.

Note: The "lock" is more symbolic than anything else as it only
      inserts a block to push to remote. The user has all
      possibilities of shooting themselves in the foot or hip as Git
      has no concept of "read-only" repos.

When cloning repos the sub-command now automatically switches to
EITHER the version (i.e. tag name) listed in the input file OR the
feature branch identified by the Jira issue ID if it exist (otherwise
to develop or master branch in that priority order).

Already cloned repos are also updated like above.

This means that by setting version field to "floating" for a repo that
does not have a feature branch, then the end result is (most likely)
that the repo will follow HEAD on develop.

Add guard to .core_preamble.bash to block multiple sourcing of it.

Bash function ensure_mode_set is made more flexible so it can handle
multiple alternative modes. Useful for command options that may be
used in multiple modes.

Make hard coded file extensions ".init" and ".repos" be defined by
global variables instead.

Update frija init command description so it aligns with how the
command is intended to work.