FLTS-4452: Add IntermediateOutputPath
IntermediateOutputPath is added to generated Directory.Build.props
file (used by MSBuild project file format for C++ projects). Unless
set when building the default value is a relative path originating
from the project root. Temporary files are placed in this directory
and unless the corresponding folder structure is created before the
build MSBuild will also issue a warning about this.

One side effect of ignoring this warning is that caching of build
results is disabled and hence the build takes longer time. Another
side effect is that the warning will trigger a failed build when
warnings are treated as errors...