FLTS-4452: Fix MSB3101 warning
MSBuild complains about the IntermediateOutputPath not existing during
build, which results in warning MSB3101 which in turn fails the build
since all warnings are treated as errors.

The solution to this problem is to ensure that this folder exists
before the build starts and this is done by simply adding a target to
the Directory.Build.props file that A) does this and B) runs before
the build.