FLTS-4452: Add support for isolated execution
Major addition in this commit is that build commands are now run in an
isolated environment (clean environment with minimal set of
environment variables). Environment to use is specified in .seci files
stored in a locale repo and referenced from the .repos file for each
repo. See documentation in frija_isolate.bash for further information.