FLTS-4452: Handle special MS Windows env variables
Windows allow environment variable names containing parentheses and
MSBuild do depend on two such variables being
set ("CommonProgramFiles(x86)" and "ProgramFiles(x86)"). Alas, Bash
does not allow such names for environment variables and due to this it
is not possible to read, set, nor export such environment variable
names even if they exist in the environment.

This causes problems when the env command is used for clearing the
environment prior to a command being executed, since it is impossible
to recreate the needed subset of the environment on Windows for
MSBuild using Bash.

As a workaround CMD.EXE is used for invoking a BAT-file that in turn
sets these two variables and then calls MSBuild.exe command. These two
special variables are assumed to be set in a .seci file but uses dummy
names that work in Bash, that is CommonProgramFiles_x86_ and
ProgramFiles_x86_ respectively. The BAT-file then do

SET CommonProgramFiles(x86)="%CommonProgramFiles_x86_%"
SET ProgramFiles(x86)="%ProgramFiles_x86_%"

and then the command is executed with its arguments. The command and
arguments are sent as parameters to the BAT-file and expanded using
'%*'.

Note that the SHIFT command in BAT-files does not work in the same way
as it does in Bash-scripts. What it does is that it only shifts which
index the positional parameters %1 through %9 point to, but it does
not shift the expansion of %* as is done in Bash. Due to this (and
other limitations of BAT-files) it is nearly impossible to do a
generic implementation that would indirectly assign the value of
foobar_fnord_ to "foobar(fnord)" and gazonk_xyzzy_ to "gazonk(xyzzy)"
if "foobar_fnord_,gazonk_xyzzy_" were sent as an argument to the
script before the actual command + arguments to execute.

Hence hardcoded workaround in the run_windows_command.bat BAT-file.