FLTS-4452: Create Fensalir environment variables
Environment variable $_FENSALIR_ROOT introduced; it points to the root
of the Fensalir repository. Existing variable $_FENSALIR_HOME is now
based on $_FENSALIR_ROOT.

Environment variable $_FENSALIR_SUPPORT introduced; it points to the
support folder within the Fensalir repository. This folder will
contain sub-folders containing "support" files, for instance GNU
Makefile fragments to be used together with GNU make when a fully
Frija-integrated repo is used together with Frija.

Note: This commit does NOT introduce the actual support folder within
the repository as Git only tracks files and not folders. The actual
folder will be added in another commit as a side effect when the first
set of support files are added to the repository.