FLTS-7671: Add GNU Make support to Frija
When running GNU Make from Frija build command there is a need to
communicate to the GNU Makefile what the name of the per-repo
generated file is. That is the name of the file that was generated by
Frija generate command.

One generic way of doing this is to set an environment variable with a
predefined name to the name of this file. A complicating factor is
that the build command is executed in a minimal environment that is
explicitly setup by Frija before calling the build command.

This dilemma is solved by letting the run_isolated() command know
about the possible existence of a global variable named
_FENSALIR_BUILD_SUPPORT. If it is set then a variable with that name
and value will be made available as an exported variable in the
isolated execution environment this function sets up via the env
command and the frija_isolate.bash script. The env command accepts a
list of name-value pairs of environment variables and their values
that should be made available in the setup environment (these are
stored in the $extraVariables[] Bash-array).