FLTS-4452: Revamp completion support
Completion implementation (and corresponding commands) have been
changed to not support any arguments, just options. The problem lies
in that it is extremely difficult to have completion support for
arguments due to that the shell built-in completion support is only
for options and not arguments.

Thus all commands have been updated so that any input file is now
specified using an option argument instead of a plain argument.