FLTS-7718: Add Make-support for sub-folders
Adds support for placing source code in multiple sub-folders of a
repo.

When building using GNU Make the Frija provided Makefile.preamble and
Makefile.reporules Makefile fragments will from now also support
including source files from multiple folders. Previously only the src
folder assumed to be located in the root of a repo was considered.

This commit enables that source code can be placed in the additional
folders listed in the variable FRIJA_EXTRA_SOURCE_FOLDERS in the
Makefile.extras file.

Note that if this variable is instead assigned in the Makefile file of
the repo, then this configuration will not be picked up when building
in parallel across all repos from a single Makefile (once that support
has been added to Frija).