FLTS-7771: Add Graphviz dependency graph generation
Add support in 'frija generate' command for generation of dependency
graph using the repos file as input. The generated dependency graph is
represented using Graphiz Dot file format; this means that the dot
tool can be used for generating an SVG (Scalable Vector Graphics)
representation of the generated graph.

Note that for instance Firefox web browser can be used for viewing SVG
files.