FLTS-7822:
When building using frija build command indirect dependencies are
never discovered as they should be. This has led us to explicitly list
all direct and indirect dependencies in the repos-file.

Makefile generated by frija generate command explicitly includes the
generated file FrijaGenerate.Makefilefragment. This in turn result
in that Makefile.reporules always selects the path where FRIJA_ID
variable is set. This is bad since the test that differentiates
between using a Global Makefile or Repo Makefile checks if FRIJA_ID is
set or not; end result is that the global-makefile-path is always
selected.

This in turn have hidden that the macro FRIJA_BUILD_DEPENDENCIES
needed to be updated to reflect that the generated per-repo file with
dependencies was renamed from <reponame>.dependencies to
FrijaGenerated.Makefilefragment.

The same goes for the call to the macro in Makefile.reporules.

  - frija-generate no longer adds an explicit include of
    FrijaGenerated.Makefilefragment in the generated per-repo
    Makefile.

  - Macro FRIJA_BUILD_DEPENDENCIES in Makefile.preamble modified to it
    now includes $(FRIJA_GENERATED_MAKEFILE_FRAGMENT) for each repo
    there are direct or discovered indirect dependencies to.

  - Makefile.reporules now calls macro FRIJA_BUILD_DEPENDENCIES with
    argument $(FRIJA_GENERATED_MAKEFILE_FRAGMENT).