FLTS-7838: Fix erroneous reporting of dirty repos
Untracked files were treated as if they caused a repo to become
'dirty', that is a version controlled file is modified and not added
to a commit.

Additionally the label "Untracked changes" was used instead of
"Untracked files" for untracked files.

The fix is to modify Bash function get_print_status so it returns
number of files with the given status. And then sum together the
reported values while ignoring reported value for untracked files. If
the sum is greater than zero then the repo is treated as if it is
dirty, otherwise not.