FLTS-7837: Add support for C extra defines
Add support for variable C_EXTRA_DEFINES that can be set in the
Makefile.extras file located in a repo that is to be built using the
GNU Makefile fragments in this repo together with frija generate and
frija build commands.

This variable enables a per-repo setting of additional defined values
to use when compiling C source code.