FLTS-7881: Add support for Fortran include files in other folders
When compiling Fortran code with GCC it does not support automatic
generation of dependency files. Current implementation in the Makefile
takes a simplified approach using sed command to extract include
statements from the Fortran source files; it does support two levels
of include files but assumes the included file is located in the same
folder as the source file.

This implementation is too simplistic and it need to search through
all folders that may contain source code and/or include files to
locate the folder containing the intended file to be included.

Instead the Makefile.preamble is modified so that it picks up the
include statements to the file and searches through all given include
paths for the requested file. If there are multiple matches then an
error message is printed (it is ambiguous which file to pick).
Otherwise the path to the file is selected and added to the list of
include paths to use when compiling the repo. The found file is then
searched for any include statements and so on.