FLTS-8040: Fix Bash 5 and Solaris problems
There are some very small differences between Bash 4.2 and Bash 5.1,
for instance when calculating number of elements in an array it is no
longer supported to provide a default value of 0 (which is
unnecessary), e.g.

${#COMP_WORDS[@]:-0}

must be replaced with

${#COMP_WORDS[@]}

Furthermore, transition from the generic global variable names
OPERATING_SYSTEM, LINUX_OS, WINDOWS_OS, and so on to the corresponding
Fensalir specific ones _FENSALIR_CURRENT_OS, _FENSALIR_LINUX,
_FENSALIR_WINDOWS, and so on is completed. This was necessary to do
due to broken Solaris support. In addition to this the variables
storing the OS version has been finalized and generalized so that in
addition to _FENSALIR_SOLARIS_VERSION there are also
_FENSALIR_LINUX_VERSION and _FENSALIR_WINDOWS_VERSION.

Furthermore, two new global variables _FENSALIR_OS_VERSION and
_FENSALIR_OS_ID are introduced. The latter reflect what may be used in
the first column of the .repos file.

Stack trace printout from subshells when they have a non-zero exit
code have been cleaned up and streamlined.