FLTS-8040: Fix Frija status breaking on Solaris
Fix usage of grep command so that it uses $_FENSALIR_GREP instead of
explict call to 'grep'. This is due to that on Solaris GNU grep is
called 'ggrep' and on Linux it is called just 'grep'. When Fensalir is
configured via users init-files a global variable called
_FENSALIR_GREP is exported. This variable is set to 'grep' on Linux
and 'ggrep' on Solaris.

Thus using this variable instead of direct calls to 'grep' solves the
grep "incompatibility" problem on Linux and Solaris.