FLTS-8108:  Add support for Ada executables
When building an Ada executable a so called "Ada Execution
Environment" must be provided; Ada programs runs in some sense in
their own universe without knowing or caring about the Operating
System that started the program. Thus Ada require this "environment"
to be present for the program; usually gnatbind provides this
environment.

But when building a library from Ada code (which is done for each repo
when using the Fensalir GNU Makefiles) gnatbind is explicitly told to
not generate the glue layer that calls the init and finalize methods
for the environment. This is handed over to the one building the
libraries to provide.

The C file included in this commit (frija_cmain.c) provides this glue
layer that initializes the Ada environment, calls the main function in
the Ada program and when it return finalizes the Ada environment.

In order to support this C file the Makefiles Makefile.preamble and
Makefile.reporules are modified to cater for this case. This means
that the Makefile will automatically insert a define definition on the
command line when compiling C code for an executable where a file
called $FRIJA_CMAIN is located in the root of the executables folder.
See documentation in the updated and included files for more
information.

Two bugs are also fixed in this commit

1. Makefile.reporules does not exit with an error for a repo that only
   contain a prebuilt binary and/or prebuilt libraries.

2. When an executable is made up of Ada source code and there are
   dependencies to other repos the list of paths to search for ALI/ADS
   files is now included on the GCC command line.