FLTS-8275: Fix race condition between cp and ln
Fix erroneous behavior where pre-compiled .so-files residing in lib
folder where copied to destination folder *after* hard links to
destination folder was created.

To understand the behavior of this bug it is very important to grok
that even though Makefiles are declarative in their nature they are
evaluated in two passes. The first pass basically evaluates all
expressions and the second pass executes rules (if necessary). This
means that this temporal aspect must always be considered when writing
and reading Makefiles, and for this bug it was the cause of the bug.

1) Rules for copying pre-compiled .so files were created when a
   library folder containing them was found in a repo. This in itself
   delays the copy operation and it will only execute if the source
   file is newer than the target file (or target file does not exist);
   that is this might execute in the *second* pass.

2) As a side effect of evaluating the Makefile expressions an explicit
   search was made in Makefile.reporules in the filesystem for
   .so-files to create hard links to. That is this is executed in
   the *first* pass.

Due to that #2 is executed before #1 resulted in that the pre-compiled
libraries were not discovered in the filesystem at the time when the
rules for creating the hard links were instantiated.

To solve this race condition the Makefile must be made aware of the
existence of the pre-compiled libraries before the rules creating hard
links to them are instantiated. This is done by maintaining a list
$FRIJA_ID values for the repos that contain pre-compiled
libraries ($FRIJA_PREBUILT_LIB_REPO_LIST).

Since discovering if a repo contain any pre-compiled libraries or not
involves a search in the filesystem, the result of this search is
cached in a per-repo variable called $(FRIJA_ID)_LIBRARY_FILES. Thus
by iterating over $FRIJA_PREBUILT_LIB_REPO_LIST it is possible to
get to the cached results. In this way it is only necessary to search
the filesystem once per repo for pre-compiled .so-files.

What is left is discerning if a repo have any dependencies to another
repo that in turn contain pre-compiled libraries. This is done by
taking the intersection between $FRIJA_PREBUILT_LIB_REPO_LIST and
$DEPENDENCY_LIST. Only when that intersection is non-empty the
corresponding lists of pre-built .so-files should be added to
$LIB_DEPS.

The implemented change thus separates the action of storing the
minimal set of per-repo library files in $LIB_DEPS from adding any
pre-built .so-files to it.

What is left to implement is handling of pre-built .a-files (static
libraries) which unfortunately is a bit more complicated as there are
two cases to take care of (repo containing pre-built static libraries
without source code and repo also containing source code). The former
means that static pre-built libraries must be included when the
executable is linked and the latter means that the pre-built static
library is automatically included when the library representing the
repo is created.