FLTS-8273: Add support for external 3PP
Add support for defining system include paths (for C and C++) for as
well as paths to precompiled libraries via SECI-files. This means that
a SECI file can represent a 3PP where the compile time interface is
defined via system include paths and the corresponding precompiled
libraries.

Please note that this commit only assumes that shared object file
format is used for the precompiled 3PP libraries.