FLTS-8273: Add support for repos without repolibs
Add support to Frija for repos that contain source code and/or
precompiled libraries but no Frija repolib is produced by the repo.

This is implemented by adding an extra column to the .repos file where
it is possible to encode variants of the supported repo kinds.
Currently the only supported repo kind-variant is 'norepolib' for the
'library' repo kind.

NOTE: THIS IS A NON-BACKWARD COMPATIBLE CHANGE TO THE .repos FILE
      FORMAT!