FLTS-8273: Update Makefiles due to quirks in GCC 3.4.6
There is a bug in GCC 3.4.6 that causes a segmentation fault when
compiling one of the Ada files in a repo where the Makefile requests
that GCC should produce Position Independent Code (PIC); this is
needed if a shared object should be created from the resulting object
file(s).

Due to this usage of the '-fPIC' flag is now only enabled when
Makefile is instructed to produce shared objects. Thus requesting that
static linkage should be used implies that no '-fPIC' flag is used and
thus the above mention segmentation fault is avoided.

Furthermore, to facilitate that it is possible to create executable
binaries where RPATH is set, and specifically set to the exact string
'$ORIGIN/../lib', absolute paths to shared objects are no longer used.
Instead -L option in combination with -l for groups of shared objects
are used. This effectively means that dynamic linker will start
searching in a folder that is relative to the folder where the
executable is located ('../lib').

To handle link order between static libraries the link commands
'--start-group' and 'end--group' are used. These link commands form a
'group' of files to repeatedly search for missing symbols until
found (simplified explanation). Effectively this means that at most
two passes are used to find all symbols during static linking. See
manual page for ld command for more information.