FLTS-8619: Fix bug related to develop-workspaces
When creating a workspace with the flag '--develop=LATEST', there was
a bug that caused frija clone to fail when changing branch, thus
exiting after cloning each repo. There was two issues that caused this
problem. First, the code for changing branch wasn't activated due to
the featureID-string being empty. Second, since the featureID-string
was empty, the glob-matching with featureID matched the first
feature-branch in the repo, thus changing to an incorrect branch.
This fix resolves this by first removing the condition on activating
the relevant code and second, by verifying that the featureID-string
is not empty before performing the glob-matching.