FLTS-8619: Add exception for .snapshot folder
The backup system creates a '.snapshot' folder in each
folder in the PWA file system. This is hidden in most
folders but is listed (with eg. 'ls') in /home. This
causes the expunge function to try to remove write-protected
temp files that are older than the backup-interval.

This prunes these folders so that they are not elig-
ible for cleaning with expunge.