FLTS-8273: Add example files
Example input files used by the Frija shell scripts as well as the
Makefile system found in this repo.

Dependencies.data is used for generating files that are imported by
the build system used. It defines things such as folders within the
repo to include or link against as well as any external libraries.

content.tools is used together with any external build tool that need
to be invoked as a part of the build process, such as Google protoc
protobuf compiler.

Makefile.extras contain additional handwritten Makefile configuration
that is automatically included by the Makefile system found in this
repo.

Makefile.buildoptions is used within a subfolder containing source
code for additional binaries that are to be built when the repo is
built.