Adapt Fensalir to Aurora domain
Adapt Fensalir to work in Aurora domain by generalizing the PWA concept;
that is it is no longer hardcoded, instead a configuration file defines
where PWA is located based on OS and the triplet country, site, and
domain.

Some files have also been renamed (as well as some functions and
variables) from ...setup... to ...init... as that is more apropriate
terminology in those cases.