Add new --trace option to frija commands
Old implementation of the run() Bash functions in Frija used the --verbose
flag to control whether the executed command line was also printed to the
terminal or not. The problem with this was that other parts of the scripts
wanted to use the --verbose flag for more verbose information than what
was possible to provide via --wordy.

Due to this, a new option called --trace, is introduced that only control
if the run() functions show the executed command line or not.

This opened up for the SECI parser functions to provide two levels of
verbose output (that can be combined): --verbose makes the read lines
from the SECI files to be echoed to the terminal (with line numbers), and
--wordy echoes the evaluated expressions to the terminal.