Add new Frija command 'frija create'
Add new command 'frija create' that might superseed the 'frija
prepare' and 'frija archive' commands that are not yet completed. The
new command is more light-weight in some ways and depend on
configuration file (volla.mapping) in the root of the Volla repos
referenced from the .repos file.

Additionally a new column is introduced in the .repos file stating
which Volla-repo is controlling each respective repo. Each
volla.mapping file states for each repo which files/folders to create
soft-links to in the Installation skeleton folder. The rules are
simply

- If source is a folder, then a soft link to the source folder is
  created with the name of the source folder in the given destination
  folder.

- If the destination folder name is '.' then this means the root of
  the Installation skeleton folder.

- If the source contain a glob expression, then this expression is
  automatically expanded and links to all the expanded files are
  placed in the destination folder.

- A column is dedicated for specifying wither the source path is
  relative to the root of the repo or relative to the Result
  folder (including type and variant). The supported values are simply
  'repo' or 'result'.

Using soft links is something that might be replaced with actual file
copying as soft links are automatically updated when new builds of the
same type are made. Using hard links is even worse since they point to
specific i-nodes and depending on the program modifying the file the
hard link points to, the link might point to an old version of the
file or the latest version. This depends on if the program modifying
the file is creating new i-nodes or not when the file content is
modified.

If and when file copying is introduced, then each invocation of 'frija
create' command will most likely create a sub-folder whos name contain
some kind of time-stamp and other meta-data.

Note: The command does not yet support creating an actual archive
file, it just links to the referenced files from the Instalation
folder (and sub-folders).