Fix minor bugs
Fix minor bugs related to TAB-completion and debugging of it.

When completing options for a command, and there is a mismatch between
option modifiers (no argument, optional argument, and mandatory
argument) for short options and long options. These must match 1:1 for
the overlapping part between short and long options. What happened was
that only the options up to this point were printed in the terminal
when completed with no message about what had happened. A detailed
warning message is now printed to the terminal.

When 'frija --debug=ON' is executed a check is made if Bash variable
COMP_TYPE is defined or not. If it is not defined a debug message
stating this was printed to the terminal in function
_frija_redraw_current_line(). Alas this function also get called as a
side-effect of the debug printout, hence an infinite loop occurs as
this variable is only defined during completion. The debug printout is
now removed from this function.

When 'frija --debug=OFF' is executed a slightly erronous message was
printed due to a copy and paste error (remants from '--debug=ON'
message had not been removed completely.